@@ -1,5 +1,5 @@
 class Admin::UsersController < Admin::ApplicationController
-  before_filter :find_user, except: [ :index ]
+  before_action :find_user, except: [ :index ]
 
   def index
     @users = User.clients
