@@ -1,8 +1,8 @@
 class OrdersController < ApplicationController
-  skip_before_filter :verify_authenticity_token, :authenticate!, only: [ :notify ]
-  before_filter :find_order, except: [ :index, :new, :create, :notify, :not_found ]
-  before_filter :redirect_pending, only: [ :new, :create ]
-  before_filter :redirect_mechanic, only: [ :update, :confirm ]
+  skip_before_action :verify_authenticity_token, :authenticate!, only: [ :notify ]
+  before_action :find_order, except: [ :index, :new, :create, :notify, :not_found ]
+  before_action :redirect_pending, only: [ :new, :create ]
+  before_action :redirect_mechanic, only: [ :update, :confirm ]
 
   helper_method :order_klass
 
