@@ -1,5 +1,5 @@
 class UserSessionsController < ApplicationController
-  skip_before_filter :authenticate!
+  skip_before_action :authenticate!
 
   def new
     new_weixin
