@@ -1,4 +1,4 @@
-class Fellowship < ActiveRecord::Base
+class Fellowship < ApplicationRecord
   belongs_to :mechanic
   belongs_to :user
 end
