@@ -1,3 +1,3 @@
-class Note < ActiveRecord::Base
+class Note < ApplicationRecord
   belongs_to :store
 end
