@@ -1,4 +1,4 @@
-class Series < ActiveRecord::Base
+class Series < ApplicationRecord
   belongs_to :brand
 
   validates_presence_of :name, :brand
